+++ controllers/resource/upserter.go
@@ -5,12 +5,14 @@ import (
 	"errors"
 	"fmt"
 	"reflect"
+	"slices"
 	"strings"
 
 	imagev1 "github.com/google/go-containerregistry/pkg/v1"
 
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
+	netv1 "k8s.io/api/networking/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -22,7 +24,11 @@ import (
 )
 
 const (
-	errMsgFailedToGetPVC = "failed to get pvc"
+	errMsgFailedToGetPVC    = "failed to get pvc"
+	k8sNginxIngressDoguName = "nginx-ingress"
+	k8sNginxStaticDoguName  = "nginx-static"
+	dependencyTypeDogu      = "dogu"
+	dependencyTypeComponent = "component"
 )
 
 var (
@@ -30,15 +36,21 @@ var (
 )
 
 type upserter struct {
-	client           k8sClient
-	generator        DoguResourceGenerator
-	exposedPortAdder exposePortAdder
+	client                 k8sClient
+	generator              DoguResourceGenerator
+	exposedPortAdder       exposePortAdder
+	networkPoliciesEnabled bool
 }
 
 // NewUpserter creates a new upserter that generates dogu resources and applies them to the cluster.
-func NewUpserter(client client.Client, generator DoguResourceGenerator) *upserter {
+func NewUpserter(client client.Client, generator DoguResourceGenerator, networkPoliciesEnabled bool) *upserter {
 	exposedPortAdder := NewDoguExposedPortHandler(client)
-	return &upserter{client: client, generator: generator, exposedPortAdder: exposedPortAdder}
+	return &upserter{
+		client:                 client,
+		generator:              generator,
+		exposedPortAdder:       exposedPortAdder,
+		networkPoliciesEnabled: networkPoliciesEnabled,
+	}
 }
 
 // UpsertDoguDeployment generates a deployment for a given dogu and applies it to the cluster.
@@ -106,6 +118,112 @@ func (u *upserter) UpsertDoguPVCs(ctx context.Context, doguResource *k8sv2.Dogu,
 	}
 
 	return nil, nil
+}
+
+// UpsertDoguNetworkPolicies generates the network policies for a dogu
+func (u *upserter) UpsertDoguNetworkPolicies(ctx context.Context, doguResource *k8sv2.Dogu, dogu *core.Dogu) error {
+	logger := log.FromContext(ctx)
+	if !u.networkPoliciesEnabled {
+		logger.Info("Do not create network policies as they are disabled by configuration")
+		return nil
+	}
+
+	var multiErr error
+	denyAllPolicy := generateDenyAllPolicy(doguResource, dogu)
+
+	if err := u.upsertNetworkPolicy(ctx, denyAllPolicy); err != nil {
+		multiErr = errors.Join(multiErr, fmt.Errorf("failed to create or update deny all rule for dogu %s: %w", dogu.GetSimpleName(), err))
+	}
+
+	var allDependencies = append(dogu.Dependencies, dogu.OptionalDependencies...)
+
+	for _, dependency := range allDependencies {
+		if dependency.Type == dependencyTypeDogu {
+			if err := u.upsertDoguDependencyNetworkPolicy(ctx, dependency.Name, doguResource, dogu); err != nil {
+				multiErr = errors.Join(multiErr, err)
+			}
+		}
+		if dependency.Type == dependencyTypeComponent {
+			if err := u.upsertComponentDependencyNetworkPolicy(ctx, dependency.Name, doguResource, dogu); err != nil {
+				multiErr = errors.Join(multiErr, err)
+			}
+		}
+	}
+
+	if err := u.deleteNonExistentDependencyPolicies(ctx, dogu, allDependencies); err != nil {
+		multiErr = errors.Join(multiErr, err)
+	}
+
+	if multiErr != nil {
+		logger := log.FromContext(ctx)
+		logger.Error(multiErr, fmt.Sprintf("failed to create some network policies for dogu %s", dogu.GetSimpleName()))
+		return multiErr
+	}
+
+	return nil
+}
+
+func (u *upserter) deleteNonExistentDependencyPolicies(ctx context.Context, dogu *core.Dogu, allDependencies []core.Dependency) error {
+	var multiErr error
+
+	currentPolicies := &netv1.NetworkPolicyList{}
+	if err := u.client.List(ctx, currentPolicies, client.MatchingLabels{"dogu.name": dogu.GetSimpleName()}); err != nil {
+		return err
+	}
+
+	for _, policy := range currentPolicies.Items {
+		// Only delete policies which rely on a dependency
+		if policy.Labels[depenendcyLabel] == "" {
+			continue
+		}
+
+		// Check if the dependency of the network policy still exists in dogu dependencies
+		stillExists := slices.ContainsFunc(allDependencies, func(dependency core.Dependency) bool {
+			return dependency.Name == policy.Labels[depenendcyLabel]
+		})
+
+		// If not existent in dogu dependency, remove the network policy
+		if !stillExists {
+			if err := u.client.Delete(ctx, &policy); err != nil {
+				multiErr = errors.Join(multiErr, err)
+			}
+		}
+	}
+
+	return multiErr
+}
+
+func (u *upserter) upsertDoguDependencyNetworkPolicy(ctx context.Context, dependencyName string, doguResource *k8sv2.Dogu, dogu *core.Dogu) error {
+	if dependencyName == k8sNginxStaticDoguName {
+		return nil
+	}
+
+	var dependencyNetworkPolicy *netv1.NetworkPolicy
+	if dependencyName == k8sNginxIngressDoguName {
+		dependencyNetworkPolicy = generateIngressNetPol(doguResource, dogu)
+	} else {
+		dependencyNetworkPolicy = generateDoguDepNetPol(doguResource, dogu, dependencyName)
+	}
+
+	if err := u.upsertNetworkPolicy(ctx, dependencyNetworkPolicy); err != nil {
+		return fmt.Errorf("failed to create or update network policy allow rule for dependency %s of dogu %s: %w", dependencyName, dogu.GetSimpleName(), err)
+	}
+
+	return nil
+}
+
+func (u *upserter) upsertComponentDependencyNetworkPolicy(ctx context.Context, dependencyName string, doguResource *k8sv2.Dogu, dogu *core.Dogu) error {
+	dependencyNetworkPolicy := generateComponentDepNetPol(doguResource, dogu, dependencyName)
+
+	if err := u.upsertNetworkPolicy(ctx, dependencyNetworkPolicy); err != nil {
+		return fmt.Errorf("failed to create or update network policy allow rule for dependency %s of dogu %s: %w", dependencyName, dogu.GetSimpleName(), err)
+	}
+
+	return nil
+}
+
+func (u *upserter) upsertNetworkPolicy(ctx context.Context, netPol *netv1.NetworkPolicy) error {
+	return u.updateOrInsert(ctx, getNetPolObjectKey(netPol), &netv1.NetworkPolicy{}, netPol)
 }
 
 func (u *upserter) upsertPVC(ctx context.Context, pvc *v1.PersistentVolumeClaim) error {